src/main/fastapi/requirement.txt
transformers
pillow
tf-keras
tensorflow-cpu
fastapi
requests
uvicorn